platform: separate the bit-banged I2C, make the external EEPROM a feature
bb_i2c joins bb_spi as a shared bit-bang driver: the board supplies the
two pins and, where they sit on a non-default SFR page, the wrapper that
selects it, so the driver itself stays page agnostic.

The 24Cxx settings store becomes a feature a board opts into with
`'nvm': '24cxx'`, the same way the wireless one works, rather than
something the sh68f89 platform hardcodes. Another board on this part can
now use flash or the on-chip block instead.

Two things fixed rather than carried across:

  * The write-protect macros were doing a read-modify-write on P4, which
    also carries three PCA outputs, so a read sampled live PWM. They go
    through the shadow now, like every other write to that port.
  * led_drive rebuilt the whole P4 byte from KB_P4_IDLE, which has write
    protect asserted, so every LED subframe re-protected the EEPROM,
    including from under a settings save. It left a half-written record
    that failed its checksum on the next boot. It now touches only the
    bits it owns.

Verified on hardware: a save made with the backlight running survives a
power cycle, twice.
